requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==48.0.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.1
fastar==0.11.0
google-auth==2.52.0
google-genai==2.0.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
markdown-it-py==4.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic-extra-types==2.11.1
pydantic-settings==2.14.1
pydantic_core==2.46.4
Pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.59.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
tenacity==9.1.4
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
